requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
